+++ src/Rocks.IntegrationTests/NonPublicMemberTests.cs
@@ -1,0 +1,29 @@
+﻿using NUnit.Framework;
+using Rocks.Exceptions;
+
+namespace Rocks.IntegrationTests;
+
+public abstract class HasProtectedMember
+{
+	protected HasProtectedMember() { }
+
+	public void CallsDoSomething() => this.DoSomething(3);
+
+	protected abstract void DoSomething(int data);
+}
+
+public static class NonPublicMemberTests
+{
+	[Test]
+	[RockCreate<HasProtectedMember>]
+	public static void ThrowExpectedException()
+	{
+		var protectedMemberExpectations = new HasProtectedMemberCreateExpectations();
+
+		var protectedMember = protectedMemberExpectations.Instance();
+
+		Assert.That(() => protectedMember.CallsDoSomething(), 
+			Throws.TypeOf<ExpectationException>()
+				.With.Message.EqualTo("No handlers were found for Void DoSomething(Int32)"));
+	}
+}